pyproject.toml
[tool.poetry]
name = "ithappens_bot"
version = "0.1.0"
description = ""
authors = ["Elisei <elisey.rav@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.5"
pydantic = "^2.3.0"
aiohttp = "^3.8.1"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.291"
mypy = "^1.5.1"
flake8 = "^6.1.0"
black = "^23.9.1"
isort = "^5.12.0"

[tool.mypy]
strict = true
plugins = "pydantic.mypy"

[tool.black]
line-length = 120
target_version = ['py311']

[tool.isort]
lines_after_imports = 2
line_length = 120
src_paths = ["bot"]

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
